Resolve some warnings
